JS-1011 Address S8907 review feedback
