+++ packages/analysis/src/jsts/rules/S8907/rule.ts
@@ -16,7 +16,7 @@
  */
 // https://sonarsource.github.io/rspec/#/rspec/S8907/javascript
 
-import type { Rule } from 'eslint';
+import type { Rule, Scope } from 'eslint';
 import type estree from 'estree';
 import ts from 'typescript';
 import { generateMeta } from '../helpers/generate-meta.js';
@@ -27,117 +27,23 @@ import {
   type RequiredParserServices,
 } from '../helpers/parser-services.js';
 import { getTypeFromTreeNode, isAny, isArrayLikeType, isStringType } from '../helpers/type.js';
+import {
+  callbackCollectionMethods,
+  methodNames,
+  methodNamesByLowerCase,
+  replacements,
+  shapeDependentMethods,
+  type Replacement,
+} from './catalog.js';
 import * as meta from './generated-meta.js';
-
-type Replacement = {
-  alternative: string;
-  minimumEcmaVersion: number;
-  reason?: string;
-  example?: string;
-};
 
 const ES5 = 5;
 const ES6 = 6;
 const ES2015 = 2015;
 const ES2016 = 2016;
-const ES2017 = 2017;
-const ES2019 = 2019;
 const ECMA_VERSION_YEAR_OFFSET = ES2015 - ES6;
 
-const ARRAY_NULLISH_REASON = 'the library handles nullish values differently from the native API';
-const COLLECTION_REASON = 'the library also accepts objects and nullish values';
-const COLLECTION_PREDICATE_REASON =
-  'the library also accepts objects, nullish values, and shorthand predicates';
 const STRING_INCLUDES_REASON = 'the library also accepts arrays, objects, and nullish values';
-const ASSIGN_REASON = 'the library handles nullish targets differently from Object.assign';
-
-/**
- * First-version catalog: keep high-confidence native API recommendations only.
- * Included methods have a clear native API target and either a local helper API
- * replacement, an object helper replacement, or an array/string receiver check.
- *
- * Removed or deferred because they were noisy in ruling:
- * - Collection aliases: all, any, collect, detect, each, forEach, foldl, foldr, inject, select.
- * - Object collection rewrites: Object.values()/Object.entries() alternatives for map/filter/reduce.
- * - String coercion helpers: endsWith, padEnd, padStart, repeat, replace, split, startsWith,
- *   toLower, toUpper, trim.
- * - Subtle semantic helpers: bind, isFinite, isNaN.
- */
-const replacements: Record<string, Replacement> = {
-  assign: cautious('Object.assign()', ASSIGN_REASON, ES2015),
-  concat: cautious('Array.prototype.concat()', ARRAY_NULLISH_REASON, ES5),
-  contains: cautious(
-    'Array.prototype.includes()',
-    'the library also accepts strings, objects, and nullish values',
-    ES2016,
-  ),
-  drop: cautious('Array.prototype.slice()', ARRAY_NULLISH_REASON, ES5, '`array.slice(n)`'),
-  dropRight: cautious(
-    'Array.prototype.slice()',
-    ARRAY_NULLISH_REASON,
-    ES5,
-    '`n === 0 ? array.slice() : array.slice(0, -n)`',
-  ),
-  entries: cautious('Object.entries()', ARRAY_NULLISH_REASON, ES2017),
-  every: cautious('Array.prototype.every()', COLLECTION_PREDICATE_REASON, ES5),
-  extendOwn: cautious('Object.assign()', ARRAY_NULLISH_REASON, ES2015),
-  fill: cautious('Array.prototype.fill()', ARRAY_NULLISH_REASON, ES2015),
-  filter: cautious('Array.prototype.filter()', COLLECTION_PREDICATE_REASON, ES5),
-  find: cautious('Array.prototype.find()', COLLECTION_PREDICATE_REASON, ES2015),
-  findIndex: cautious('Array.prototype.findIndex()', COLLECTION_PREDICATE_REASON, ES2015),
-  flatten: cautious(
-    'Array.prototype.flat()',
-    'the library handles nullish values differently from the native API',
-    ES2019,
-  ),
-  includes: cautious(
-    'Array.prototype.includes()',
-    'the library also accepts strings, objects, and nullish values',
-    ES2016,
-  ),
-  indexOf: cautious('Array.prototype.indexOf()', ARRAY_NULLISH_REASON, ES5),
-  isArray: direct('Array.isArray()', ES5),
-  isInteger: direct('Number.isInteger()', ES2015),
-  join: cautious('Array.prototype.join()', ARRAY_NULLISH_REASON, ES5),
-  keys: cautious('Object.keys()', ARRAY_NULLISH_REASON, ES5),
-  lastIndexOf: cautious('Array.prototype.lastIndexOf()', ARRAY_NULLISH_REASON, ES5),
-  map: cautious('Array.prototype.map()', COLLECTION_REASON, ES5),
-  pairs: cautious('Object.entries()', ARRAY_NULLISH_REASON, ES2017),
-  reduce: cautious('Array.prototype.reduce()', COLLECTION_PREDICATE_REASON, ES5),
-  reduceRight: cautious('Array.prototype.reduceRight()', COLLECTION_PREDICATE_REASON, ES5),
-  reverse: cautious('Array.prototype.reverse()', ARRAY_NULLISH_REASON, ES5),
-  slice: cautious('Array.prototype.slice()', ARRAY_NULLISH_REASON, ES5),
-  some: cautious('Array.prototype.some()', COLLECTION_PREDICATE_REASON, ES5),
-  takeRight: cautious(
-    'Array.prototype.slice()',
-    ARRAY_NULLISH_REASON,
-    ES5,
-    '`n === 0 ? [] : array.slice(-n)`',
-  ),
-  toPairs: cautious('Object.entries()', ARRAY_NULLISH_REASON, ES2017),
-  uniq: cautious('Set', ARRAY_NULLISH_REASON, ES2015, '`[...new Set(values)]`'),
-  values: cautious('Object.values()', ARRAY_NULLISH_REASON, ES2017),
-};
-
-const methodNames = new Set(Object.keys(replacements));
-const methodNamesByLowerCase = new Map(
-  Object.keys(replacements).map(method => [method.toLowerCase(), method]),
-);
-const shapeDependentMethods = new Set([
-  'contains',
-  'every',
-  'filter',
-  'find',
-  'findIndex',
-  'includes',
-  'map',
-  'reduce',
-  'reduceRight',
-  'some',
-]);
-const callbackCollectionMethods = new Set(
-  [...shapeDependentMethods].filter(method => method !== 'contains' && method !== 'includes'),
-);
 
 type CollectionShape = 'array' | 'string' | 'unknown';
 
@@ -197,19 +103,6 @@ export const rule: Rule.RuleModule = {
     };
   },
 };
-
-function direct(alternative: string, minimumEcmaVersion: number): Replacement {
-  return { alternative, minimumEcmaVersion };
-}
-
-function cautious(
-  alternative: string,
-  reason: string,
-  minimumEcmaVersion: number,
-  example?: string,
-): Replacement {
-  return { alternative, reason, minimumEcmaVersion, example };
-}
 
 function getMessageId(replacement: Replacement): 'direct' | 'cautious' | 'cautiousWithExample' {
   if (replacement.reason === undefined) {
@@ -286,15 +179,14 @@ function hasUnsupportedStringSearchValue(
 function getConstantExpression(
   node: estree.Expression,
   context: Rule.RuleContext,
-  visited = new Set<string>(),
+  visited = new Set<Scope.Variable>(),
 ): estree.Expression {
-  if (!isIdentifier(node) || visited.has(node.name)) {
+  if (!isIdentifier(node)) {
     return node;
   }
-  visited.add(node.name);
 
   const variable = getVariableFromName(context, node.name, node);
-  if (variable?.defs.length !== 1) {
+  if (!variable || visited.has(variable) || variable.defs.length !== 1) {
     return node;
   }
 
@@ -309,6 +201,7 @@ function getConstantExpression(
     return node;
   }
 
+  visited.add(variable);
   return getConstantExpression(init, context, visited);
 }
 
@@ -333,7 +226,11 @@ function getShapeSpecificReplacement(
     return replacement;
   }
   if (shape === 'string' && (method === 'contains' || method === 'includes')) {
-    return cautious('String.prototype.includes()', STRING_INCLUDES_REASON, ES2016);
+    return {
+      alternative: 'String.prototype.includes()',
+      minimumEcmaVersion: ES2016,
+      reason: STRING_INCLUDES_REASON,
+    };
   }
   return null;
 }
